LayoutTests/media/modern-media-controls/status-label/status-label-ios-expected.txt

Testing StatusText properties when used in IOSInlineMediaControls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body.contains(mediaControls.statusLabel.element) became true
PASS style.top is "14.5px"
PASS style.color is "rgb(0, 0, 0)"

PASS successfullyParsed is true

TEST COMPLETE
